remove arm copyright from newly added files update formatting made microbit specific changes to be board configurable via yaml files
