+++ source/board/kl27z_microbit_bl.c
@@ -3,7 +3,6 @@
  * @brief   board ID and meta-data for the hardware interface circuit (HIC) based on the NXP KL27Z
  *
  * DAPLink Interface Firmware
- * Copyright (c) 2009-2019, ARM Limited, All Rights Reserved
  * Copyright 2020 NXP
  * SPDX-License-Identifier: Apache-2.0
  *
