+++ source/board/microbitv2/flexio_pwm.h
@@ -3,7 +3,6 @@
  * @brief
  *
  * DAPLink Interface Firmware
- * Copyright (c) 2009-2016, ARM Limited, All Rights Reserved
  * Copyright 2020 NXP
  * SPDX-License-Identifier: Apache-2.0
  *
