+++ source/hic_hal/freescale/kl27z/usb_config.c
@@ -68,7 +68,12 @@
 //       <i> when the device is fully operational (bMaxPower)
 //   </h>
 #define USBD_CFGDESC_BMATTRIBUTES   0x80
-#define USBD_CFGDESC_BMAXPOWER      0x96    // Temporal modification to 300mA for microbit
+
+#if !defined(BOARD_USB_BMAXPOWER)
+#define USBD_CFGDESC_BMAXPOWER      0xFA
+#else
+#define USBD_CFGDESC_BMAXPOWER      BOARD_USB_BMAXPOWER
+#endif
 
 //   <h> String Settings
 //     <i> These settings affect String Descriptor
